+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- changed agent to always run connect with clean session
+
+### Fixed
+
+- agent simplification should fix subscription issues when broker gets restarted
+
 ## [3.0.3] - Feb 09 2022
 
 ### Fixed